Create dmi_util.c and move *_util.c to hardinfo/
These stand alone interfaces and are used across modules, so just
make them available to any module by default.
