+++ hardinfo/cpu_util.c
