+++ hardinfo/cpubits.c
