+++ hardinfo/dt_util.c
